pyproject.toml
[tool.poetry]
name = "wichtler"
version = "0.1.0"
description = ""
authors = ["Timon Schmelzer <timon.schmelzer@tu-dortmund.de>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
